app/src/main/java/com/example/simpleparadox/listycity/CityList.java
package com.example.simpleparadox.listycity;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

/**
 * This is a class that keeps track of a list of city objects
 */
public class CityList {
    private List<City> cities = new ArrayList<>();

    /**
     * This adds a city to the list if the city does not exist
     * @param city
     *      This is a candidate city to add
     */
    public void add(City city) {
        if (cities.contains(city)) {
            throw new IllegalArgumentException();
        }
        cities.add(city);
    }

    /**
     * This returns a sorted list of cities
     * @return
     *      Return the sorted list
     */
    public List<City> getCities() {
        List<City> list = cities;
        Collections.sort(list);
        return list;
    }








    /**
     * delete a city if it's in the list or throw an error
     * @param city
     *      will check and delete a city if it's already there
     */
    public void deleteCity(City city) {
        if(cities.contains(city)) {
            cities.remove(city);
        }
        else {
            throw new IllegalArgumentException();
        }
    }




    /**
     * it will return the number of city is cities list
     * @return
     *      number of city in the list
     */
    public int countNumberOfCity() {
        return cities.size();
    }


}
